requirements-dev.txt
# 开发/测试依赖（本地环境可选安装）
# 在 Unreal 环境下不强制安装，仅用于运行纯 Python 测试
pillow>=9.0.0
numpy>=1.24.0
pytest>=7.0.0
